debugcc: allow NULL primary mux
For sc8280xp/mc_cc we do not have a correct primary mux value. Currently
the value is not used at all, since the GCC debug mux is not used for
mccc measurement. Instead of programming the dummy value, skip
programming it at all.

TODO: consider reverting this commit if we find a necessity to use
debug mux for mccc or other DDR clocks.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>