README: Add a README.md
Add a README with a reminder to switch off CONFIG_IO_STRICT_DEVMEM. I
had debugcc working on 8939 and then tried it on 8280xp. It didn't work
but I remembered that there was a CONFIG option that needed to change.

It took a few minutes of menuconfig and Google to recall what that
config option was.

For future reference, lets just document it and whatever else in a
README.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>