meson: add option to build debugcc as a dynamically linked binary
debugcc is currently linked statically using -static and -static-libgcc.
This can cause issues for distributions that prefer dynamically linked
binaries (e.g. larger binaries and duplicated runtime code which could
otherwise be shared).

Add a Meson boolean option `dynamic-linking` which when enabled removes
the explicit static link flags so that debugcc is built as a dynamically
linked binary using the default toolchain behaviour. The default for this
option is false so that existing builds continue to get a statically
linked binary unless they opt in to dynamic linking.

Signed-off-by: Christopher Obbard <christopher.obbard@linaro.org>